# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "checkedspot",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9191",
            "description": "Optional server description, e.g. Main (production) server"
        },
        {
            "url": "{protocol}://api.example.com/v1",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                }
            },
            "description": "Optional server description, e.g. Main (production) server"
        },
        {
            "url": "{protocol}://staging-api.example.com",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                }
            },
            "description": "Optional server description, e.g. staging server for testing"
        }
    ],
...
    "paths": {
...
        "/project/expenditure": {
            "get": {
                "tags": [
                    "functionTest"
                ],
                "summary": "Returns Expenditure data(Array of Objects) for a particular Project",
                "description": "To Return Expenditure Data(Array of Objects), this endpoint Accepts <b>projectId</b> in <b>query</b> params",
                "operationId": "getExpenditure",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "required": true,
                        "description": "Takes <b>projectId</b> in query params to return Expenditure Data(Array of Objects)",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                        }
                    }
                ],

                "responses" : {
                    "200" :  {
                        "description" : "Array of Objects will be return to the user of this endpoint",
                        "content" : {
                            "application/json" : {
                                "type" : "array",
                                "items": {
                                    "type": "object",
                                    "description": "individual expenses"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "projectId not Found"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            },

            "post": {

            },

            "put": {
                "tags": [
                    "projects"
                ],
                "summary": "Returns Updated Expenditure Node belongs to the particular project",
                "description": "Takes an Object contains data for new expenditure which contains <b>projectId</b> and other expenditure data in <b>Body</b> and Updates the expenditures property of the Expenditure Node with the new expenditure object",
                "operationId": "updateExpenditure",
                "requestBody" : {
                    "content" : {
                        "application/x-www-form-urlencoded" : {
                            "schema" : {
                                "type": "object",
                                "properties" : {
                                    "projectId" : {
                                        "type" : "integer",
                                        "format" : "int32",
                                        "minimum" : 0,
                                        "description": "projectId"
                                    },
                                    "type" : {
                                        "type" : "string",
                                        "description" : "pass type of expenditure, Example: construction, garden maintanance, labour charge etc."
                                    },
                                    "spentOn" : {
                                        "type" : "string",
                                        "format" : "date-time",
                                        "description" : "pass the stringified date-time(2017-07-21T17:32:28Z) on which the transaction was done"
                                    },
                                    "spentBy" : {
                                        "type" : "string",
                                        "description" : "pass the name of the person who had spend"
                                    },
                                    "currency" : {
                                        "type" : "string",
                                        "description" : "pass currency"
                                    },
                                    "amount" : {
                                        "type" : "number",
                                        "format" : "float",
                                        "minimum" : 0,
                                        "description": "pass the spend amount corresponding to currency"
                                    },
                                    "description" : {
                                        "type" : "string",
                                        "description" : "pass expenditure's description"
                                    },
                                    "spentUsing" : {
                                        "type" : "string",
                                        "description" : "pass the method of payment like, GPay, PhonePay, Net Banking, NEFT, etc.."
                                    },
                                    "attachment" : {
                                        "type" : "string",
                                        "format" : "url",
                                        "description" : "pass the url of the payment receipt"
                                    }
                                },
                                "required" :["projectId", "spentOn"]
                            }
                        }
                    }
                },

                "responses" : {
                    "200" :  {
                        "description" : "As a response, this endpoint update <b>Expenditure Node's</b> expenditures property and return an Array of Object having the individual expenditure data in each object",
                        "content" : {
                            "application/json" : {
                                "type" : "array",
                                "items" : {
                                    "type" : "object",
                                    "description" : "This Object contains details individual expenses details"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "projectId not Found"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            },

            "delete": {
                "tags": [
                    "projects"
                ],
                "summary": "Returns confirmation message for the deletion of the <b>Expenditure Node</b>",
                "description": "To delete <b>Expenditure Node</b>, this endpoint Accepts <b>projectId</b> in <b>query</b> params",
                "operationId": "deleteExpenditure",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "required": true,
                        "description": "Takes <b>projectId</b> in query params to delete the Expenditure Node",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                        }
                    }
                ],

                "responses" : {
                    "200" :  {
                        "description" : "As a response, this endpoint return confirmation message for the deletion of the Expenditure Node for the particular projectId",
                        "content" : {
                            "application/json" : {
                                "type" : "string",
                                "description" : "Message for the confirmation of the deletion of Node"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "projectId not Found"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
...
    },
...
}